Revert "Recommendation engine work in progress"
This reverts commit 7f179867ba1e9251f19439f5a7ee82cfa5b36261.

+++ WorkoutTracker.Application/Exercises/ExerciseAmountRecommendation.cs
@@ -9,7 +9,7 @@ namespace WorkoutTracker.Application.Exercises
         public int ExerciseId { get; set; }
         public decimal ResistanceAmount { get; set; }
         public string ResistanceMakeup { get; set; }
-        public ushort Reps { get; set; }
+        public int Reps { get; set; }
         public string Reason { get; set; }
     }
 }
